refactor enrollment
