--- src/SiteEnrollment/deprecated/components/PhoneInput.js
@@ -1,20 +1,0 @@
-// Dependency Imports
-import React from 'react';
-import TextField from '@material-ui/core/TextField';
-
-// Default function
-const phoneInput = (props, ref) => {
-  return (
-    <TextField
-      {...props}
-      inputRef={ref}
-      autoComplete="new-phone"
-      fullWidth
-      size="small"
-      label="Phone Number"
-      name="phone"
-    />
-  );
-};
-
-export default React.forwardRef(phoneInput);
