--- src/SiteEnrollment/newSiteInfo.scss
@@ -1,19 +1,0 @@
-.cardCloseIcon {
-  overflow: hidden;
-  width: 25px;
-  height: 25px;
-  position: absolute;
-  transform: translate(-50%, -50%);
-  /* left: 0; */
-  display: flex;
-  box-sizing: border-box;
-  border-radius: 50%;
-  transition: all 0.2s ease-in;
-  padding: 0.1em;
-  align-items: center;
-  justify-content: center;
-  &:hover {
-    background: #eee;
-    cursor: pointer;
-  }
-}
