+++ src/SiteInformation/Enrollment/SiteEnrollment.js
@@ -1,21 +1,20 @@
 // Dependency Imports
 import { Button, Grid, Snackbar, Typography } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
-import Axios from 'axios';
 import React, { useState, useEffect } from 'react';
-import { useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
+import { setEnrollNewSite, setSavedData } from '../../Store/actions';
 
 // Local Imports
-// import { Context } from '../Store/Store';
-import { apiPassword, apiURL, apiUsername } from '../utils/api_secret';
-import { bannedRoles } from '../utils/constants';
-import { BannedRoleMessage, CustomLoader } from '../utils/CustomComponents';
-import EnrollNewSite from './EnrollNewSite';
+import { bannedRoles } from '../../utils/constants';
+import { BannedRoleMessage, CustomLoader } from '../../utils/CustomComponents';
+import EnrollNewSite from '../Shared/EnrollNewSite';
+import { getStats } from '../Shared/functions';
 
 // Default function
-const SiteEnrollment = (props) => {
-  // const [state] = useContext(Context);
-  const userInfo = useSelector((state) => state.userInfo);
+const SiteEnrollment = () => {
+  const dispatch = useDispatch();
+
   const [totalSitesEnrolled, setTotalSitesEnrolled] = useState(0);
   const [stateSitesEnrolled, setStateSitesEnrolled] = useState(0);
   const [showStateSpecificSites, setShowStateSpecificSites] = useState(false);
@@ -23,8 +22,11 @@ const SiteEnrollment = (props) => {
   const [bannedRolesCheckMessage, setBannedRolesCheckMessage] = useState(
     'Checking your permissions..',
   );
-  const [enrollNewSite, setEnrollNewSite] = useState(false);
-  const [savedData, setSavedData] = useState(false);
+
+  const enrollNewSite = useSelector((state) => state.enrollmentData.enrollNewSite);
+  const savedData = useSelector((state) => state.enrollmentData.savedData);
+  const userInfo = useSelector((state) => state.userInfo);
+
   const [fetchingStats, setFetchingStats] = useState(true);
   useEffect(() => {
     if (userInfo.state) {
@@ -56,7 +58,7 @@ const SiteEnrollment = (props) => {
       <Grid item xs={12}>
         <Button
           variant="contained"
-          onClick={() => setEnrollNewSite(!enrollNewSite)}
+          onClick={() => dispatch(setEnrollNewSite(!enrollNewSite))}
           color={enrollNewSite ? 'secondary' : 'primary'}
         >
           {enrollNewSite ? 'Cancel' : 'Enroll Site'}
@@ -64,13 +66,7 @@ const SiteEnrollment = (props) => {
       </Grid>
 
       {enrollNewSite ? (
-        <EnrollNewSite
-          enrollNewSite={enrollNewSite}
-          setEnrollNewSite={setEnrollNewSite}
-          saveData={savedData}
-          setSaveData={setSavedData}
-          {...props}
-        />
+        <EnrollNewSite />
       ) : (
         <Grid item xs={12}>
           {!fetchingStats ? (
@@ -95,9 +91,9 @@ const SiteEnrollment = (props) => {
         anchorOrigin={{ horizontal: 'center', vertical: 'bottom' }}
         open={savedData}
         autoHideDuration={6000}
-        onClose={() => setSavedData(false)}
+        onClose={() => dispatch(setSavedData(false))}
       >
-        <Alert onClose={() => setSavedData(false)} severity="success">
+        <Alert onClose={() => dispatch(setSavedData(false))} severity="success">
           Site successfully enrolled!
         </Alert>
       </Snackbar>
@@ -108,24 +104,3 @@ const SiteEnrollment = (props) => {
 };
 
 export default SiteEnrollment;
-
-// Helper functions
-const getStats = async (state) => {
-  let records = await fetchStats(state);
-
-  let data = records.data.data;
-
-  return data;
-};
-
-const fetchStats = async (state) => {
-  return await Axios.get(`${apiURL}/api/total/sites/${state}`, {
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    auth: {
-      username: apiUsername,
-      password: apiPassword,
-    },
-  });
-};
