+++ src/Store/actions/allDataTableActions.js
@@ -1,24 +1,3 @@
-export const setValuesEdited = (valuesEdited) => {
-  return {
-    type: 'SET_VALUES_EDITED',
-    payload: valuesEdited,
-  };
-};
-
-export const setEditModalOpen = (editModalOpen) => {
-  return {
-    type: 'SET_EDIT_MODAL_OPEN',
-    payload: editModalOpen,
-  };
-};
-
-export const setEditModalData = (editModalData) => {
-  return {
-    type: 'SET_EDIT_MODAL_DATA',
-    payload: editModalData,
-  };
-};
-
 export const setReassignSiteModalOpen = (reassignSiteModalOpen) => {
   return {
     type: 'SET_REASSIGN_SITE_MODAL_OPEN',
