+++ src/Store/actions/index.js
@@ -1,3 +1,5 @@
 export * from './allDataTableActions';
+export * from './enrollmentActions';
 export * from './formsDataActions';
+export * from './sharedSiteInfoActions';
 export * from './userInfoActions';
