+++ src/Store/reducers/allDataTableReducer.js
@@ -13,14 +13,8 @@ const defaultState = {
 };
 
 // Default function
-export const allDataTableReducer = (state = defaultState, action) => {
+export const allDataTableReducer = (state = defaultState, action = {}) => {
   switch (action.type) {
-    case 'SET_VALUES_EDITED':
-      return setValuesEdited(state, action);
-    case 'SET_EDIT_MODAL_OPEN':
-      return setEditModalOpen(state, action);
-    case 'SET_EDIT_MODAL_DATA':
-      return setEditModalData(state, action);
     case 'SET_REASSIGN_SITE_MODAL_OPEN':
       return setReassignSiteModalOpen(state, action);
     case 'SET_REASSIGN_SITE_MODAL_DATA':
@@ -42,60 +36,6 @@ export const allDataTableReducer = (state = defaultState, action) => {
   }
 };
 
-// const setValuesEdited = (state, action) => {
-//   state.tableData.valuesEdited = action.payload;
-// };
-// const setEditModalOpen = (state, action) => {
-//   state.tableData.editModalOpen = action.payload;
-// };
-// const setEditModalData = (state, action) => {
-//   state.tableData.editModalData = action.payload;
-// };
-// const setReassignSiteModalOpen = (state, action) => {
-//   state.tableData.reassignSiteModalOpen = action.payload;
-// };
-// const setReassignSiteModalData = (state, action) => {
-//   state.tableData.reassignSiteModalData = action.payload;
-// };
-// const setUnenrollOpen = (state, action) => {
-//   state.tableData.unenrollOpen = action.payload;
-// };
-// const setUnenrollRowData = (state, action) => {
-//   state.tableData.unenrollRowData = action.payload;
-// };
-// const setShowNewIssueDialog = (state, action) => {
-//   state.tableData.showNewIssueDialog = action.payload;
-// };
-// const setNewIssueData = (state, action) => {
-//   state.tableData.newIssueData = action.payload;
-// };
-// const setMapModalOpen = (state, action) => {
-//   state.tableData.mapModalOpen = action.payload;
-// };
-// const setMapModalData = (state, action) => {
-//   state.tableData.mapModalData = action.payload;
-// };
-
-const setValuesEdited = (state, action) => {
-  return {
-    ...state,
-    valuesEdited: action.payload,
-  };
-};
-const setEditModalOpen = (state, action) => {
-  console.log('modal');
-  return {
-    ...state,
-    editModalOpen: action.payload,
-  };
-};
-const setEditModalData = (state, action) => {
-  //   console.log(action.payload.valuesEdited);
-  return {
-    ...state,
-    editModalData: action.payload,
-  };
-};
 const setReassignSiteModalOpen = (state, action) => {
   return {
     ...state,
@@ -121,7 +61,6 @@ const setUnenrollRowData = (state, action) => {
   };
 };
 const setShowNewIssueDialog = (state, action) => {
-  console.log('first');
   return {
     ...state,
     showNewIssueDialog: action.payload,
