--- src/Table/AllDataTable.md
@@ -1,1 +1,0 @@
-Includes various sub-components
