+++ src/utils/SharedFunctions.js
@@ -105,7 +105,7 @@ export const callAzureFunction = async (data, endpoint, method, getTokenSilently
 
   try {
     correctionsApiResponse = await fetch(
-      `https://correctionsapi.azurewebsites.net/api/${endpoint}`,
+      `https://devcorrectionsapi.azurewebsites.net/api/${endpoint}`,
       options,
     );
   } catch (err) {
