+++ src/utils/constants.js
@@ -3,7 +3,7 @@ import Axios from 'axios';
 import qs from 'qs';
 
 // Local Imports
-import { apiUsername, apiPassword, apiURL, onfarmAPI } from './api_secret';
+import { apiUsername, apiPassword, apiURL } from './api_secret';
 import { apiCorsUrl } from '../Devices/hologramConstants';
 
 // anyone with these roles are not supposed to view anything !
@@ -134,17 +134,6 @@ export const statesHash = {
   WV: 'West Virginia',
   WI: 'Wisconsin',
   WY: 'Wyoming',
-};
-
-export const fetchGrowerByLastName = async (query, apiKey) => {
-  const response = await fetch(`${onfarmAPI}/producers?search=${query}`, {
-    headers: {
-      'x-api-key': apiKey,
-    },
-  });
-
-  const data = await response.json();
-  return data;
 };
 
 // export const ucFirst = (str = "") => {
